gemini response parse function added to utility folder, ide (codemirror) now displays the default function name in the relevant language the user chose
